Initial conversion of mesh into two test sets
