+++ Source/AssetGenerator.csproj
@@ -82,7 +82,8 @@
     <Compile Include="Tests\Material_SpecularGlossiness.cs" />
     <Compile Include="Tests\Material_MetallicRoughness.cs" />
     <Compile Include="Tests\Compatibility.cs" />
-    <Compile Include="Tests\Mesh.cs" />
+    <Compile Include="Tests\Mesh_Primitives.cs" />
+    <Compile Include="Tests\Mesh_Indices.cs" />
     <Compile Include="Tests\Primitive_Attribute.cs" />
     <Compile Include="Tests\Texture_Sampler.cs" />
     <Compile Include="Test.cs" />
@@ -97,7 +98,8 @@
     <EmbeddedResource Include="Tests\Texture_Sampler.md" />
     <EmbeddedResource Include="Tests\Material_SpecularGlossiness.md" />
     <EmbeddedResource Include="Tests\Compatibility.md" />
-    <EmbeddedResource Include="Tests\Mesh.md" />
+    <EmbeddedResource Include="Tests\Mesh_Indices.md" />
+    <EmbeddedResource Include="Tests\Mesh_Primitives.md" />
   </ItemGroup>
   <ItemGroup>
     <EmbeddedResource Include="Tests\panel_baseColor.png" />
