+++ Source/Tests/Mesh_Indices.md
@@ -2,6 +2,10 @@ No attributes are explicitly set in every model generated by this test, other th
 
 All values of Byte, Short, and Int are unsigned.  
  
+ Indices |
+:---: |
+<img src="./Icon_Indices.png" height="72" width="72" align="middle">
+
 The following table shows the properties that are set for a given model.  
 
 ~~Table~~
