+++ Source/Tests/Mesh_Primitives.md
@@ -1,9 +1,5 @@
 No attributes are explicitly set in every model generated by this test, other than those in the base model.  
 
-If both Vertex UV 0 and 1 are set on a model, then all of the textures will use Vertex UV 1.
-
-All values of Byte and Short are normalized unsigned.
-
 Indices for Primitive 1 (Left) | Indices for Primitive 2 (Right)
 :---: | :---:
 <img src="./Icon_Indices_Primitive1.png" height="72" width="72" align="middle"> | <img src="./Icon_Indices_Primitive2.png" height="72" width="72" align="middle">
